+++ install.sh
@@ -68,6 +68,7 @@ config = {
         "top": "$TOP_CAMERA",
         "bottom": "$BOTTOM_CAMERA",
         "camera_connect_timeout_sec": 5,
+        "camera_frame_timeout_sec": 10,
         "camera_reconnect_delay_sec": 2,
         "ffprobe_timeout": 5,
     },